requirements.txt
pulsar-client>=2.8.1
pytest~=6.2.5